requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# requirements.txt  –  Salary Prediction MLOps Pipeline
# ══════════════════════════════════════════════════════════════════════════════

# ── Core data science ─────────────────────────────────────────────────────────
numpy>=1.24.0,<2.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ── ML / ensemble ─────────────────────────────────────────────────────────────
xgboost>=2.0.0
lightgbm>=4.0.0

# ── DVC (pipeline orchestration) ──────────────────────────────────────────────
dvc>=3.0.0

# ── Experiment tracking ───────────────────────────────────────────────────────
wandb>=0.16.0

# ── Kaggle dataset download ───────────────────────────────────────────────────
kaggle>=1.6.0

# ── Configuration & env ───────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0

# ── API serving ───────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# ── Reporting & utilities ─────────────────────────────────────────────────────
joblib>=1.3.0
tqdm>=4.66.0
